# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PRISM Engine API",
    "description": "Multi-dimensional STEAM career guidance: psychometric + financial vectorisation, financial constraint solver, parent-student conflict index, market blending, SWOT and roadmaps.\n\nAll /api/v1 responses use the envelope {success, data, error, meta}. MOCK_MODE is currently **True**.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/families": {
      "post": {
        "tags": [
          "family"
        ],
        "summary": "Create Family",
        "operationId": "create_family_api_v1_families_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_FamilyOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/families/me": {
      "get": {
        "tags": [
          "family"
        ],
        "summary": "My Family",
        "operationId": "my_family_api_v1_families_me_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_FamilyOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/families/invites": {
      "post": {
        "tags": [
          "family"
        ],
        "summary": "Create Invite",
        "operationId": "create_invite_api_v1_families_invites_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_InviteOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/families/join": {
      "post": {
        "tags": [
          "family"
        ],
        "summary": "Join Family",
        "operationId": "join_family_api_v1_families_join_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinFamilyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_FamilyOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/families/{family_id}/finance": {
      "get": {
        "tags": [
          "family"
        ],
        "summary": "Get Finance",
        "description": "Parents get raw figures (FamilyFinanceOut); students get a non-numeric FamilyFinanceSummary.",
        "operationId": "get_finance_api_v1_families__family_id__finance_get",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Union_FamilyFinanceOut__FamilyFinanceSummary__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "family"
        ],
        "summary": "Put Finance",
        "operationId": "put_finance_api_v1_families__family_id__finance_put",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamilyFinanceIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_FamilyFinanceOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/families/{family_id}/preferences": {
      "get": {
        "tags": [
          "family"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_api_v1_families__family_id__preferences_get",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ParentPreferencesOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "family"
        ],
        "summary": "Put Preferences",
        "operationId": "put_preferences_api_v1_families__family_id__preferences_put",
        "parameters": [
          {
            "name": "family_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Family Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParentPreferencesIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ParentPreferencesOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/consents": {
      "post": {
        "tags": [
          "family"
        ],
        "summary": "Create Consent",
        "operationId": "create_consent_api_v1_consents_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ConsentOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "family"
        ],
        "summary": "List Consents",
        "operationId": "list_consents_api_v1_consents_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer <access_token>",
              "title": "Authorization"
            },
            "description": "Bearer <access_token>"
          },
          {
            "name": "x-mock-role",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Role"
                },
                {
                  "type": "null"
                }
              ],
              "description": "MOCK_MODE only: impersonate a role to test role-specific views",
              "title": "X-Mock-Role"
            },
            "description": "MOCK_MODE only: impersonate a role to test role-specific views"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_ConsentOut__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConsentIn": {
        "properties": {
          "consent_type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "subject_user_id": {
            "type": "string",
            "title": "Subject User Id",
            "description": "Whose data the consent covers"
          },
          "granted": {
            "type": "boolean",
            "title": "Granted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "consent_type",
          "subject_user_id",
          "granted"
        ],
        "title": "ConsentIn"
      },
...
      "Envelope_ConsentOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[ConsentOut]"
      },
      "Envelope_FamilyFinanceOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FamilyFinanceOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[FamilyFinanceOut]"
      },
      "Envelope_FamilyOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FamilyOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[FamilyOut]"
      },
...
      "Envelope_InviteOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InviteOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[InviteOut]"
      },
...
      "Envelope_ParentPreferencesOut_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParentPreferencesOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[ParentPreferencesOut]"
      },
...
      "Envelope_Union_FamilyFinanceOut__FamilyFinanceSummary__": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FamilyFinanceOut"
              },
              {
                "$ref": "#/components/schemas/FamilyFinanceSummary"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "type": "object",
        "required": [
          "success",
          "meta"
        ],
        "title": "Envelope[Union[FamilyFinanceOut, FamilyFinanceSummary]]"
      },
...
      "FamilyFinanceIn": {
        "properties": {
          "income_band": {
            "$ref": "#/components/schemas/IncomeBand"
          },
          "annual_income": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0,
                "description": "Amount in Indian Rupees (whole rupees)"
              },
              {
                "type": "null"
              }
            ],
            "title": "Annual Income",
            "description": "Optional exact figure; band midpoint used if absent"
          },
          "income_growth_rate": {
            "type": "number",
            "maximum": 0.3,
            "minimum": -0.2,
            "title": "Income Growth Rate",
            "default": 0.05
          },
          "allocatable_savings": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Allocatable Savings",
            "description": "Amount in Indian Rupees (whole rupees)"
          },
          "existing_debt_emi": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Existing Debt Emi",
            "description": "Monthly EMI already being paid",
            "default": 0
          },
          "dependents": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Dependents",
            "default": 1
          },
          "max_affordable_emi": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Max Affordable Emi",
            "description": "Max monthly education-loan EMI the family can bear"
          },
          "loan_tolerance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Loan Tolerance",
            "description": "Normalised to [0, 1]"
          },
          "risk_appetite": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Risk Appetite",
            "description": "Normalised to [0, 1]"
          },
          "relocation_willingness": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Relocation Willingness",
            "description": "Normalised to [0, 1]"
          },
          "abroad_willingness": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Abroad Willingness",
            "description": "Normalised to [0, 1]"
          },
          "time_to_earn_years": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 2.0,
            "title": "Time To Earn Years",
            "description": "Years until the family expects the child to earn"
          },
          "prestige_vs_stability": {
            "$ref": "#/components/schemas/PrestigeStability",
            "default": "balanced"
          },
          "preferred_regions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Preferred Regions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "income_band",
          "allocatable_savings",
          "max_affordable_emi",
          "loan_tolerance",
          "risk_appetite",
          "relocation_willingness",
          "abroad_willingness",
          "time_to_earn_years"
        ],
        "title": "FamilyFinanceIn",
        "description": "Parent-side financial vector. Raw values are visible to parents only."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncomeBand": {
        "type": "string",
        "enum": [
          "below_3l",
          "3l_6l",
          "6l_10l",
          "10l_20l",
          "20l_50l",
          "above_50l"
        ],
        "title": "IncomeBand"
      },
...
      "JoinFamilyRequest": {
        "properties": {
          "invite_code": {
            "type": "string",
            "pattern": "^[A-Z0-9]{8}$",
            "title": "Invite Code"
          },
          "relation": {
            "type": "string",
            "title": "Relation",
            "examples": [
              "mother",
              "father",
              "guardian",
              "self"
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "invite_code",
          "relation"
        ],
        "title": "JoinFamilyRequest"
      },
...
      "ParentPreferencesIn": {
        "properties": {
          "preferences": {
            "items": {
              "$ref": "#/components/schemas/RankedPreference"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Preferences"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "preferences"
        ],
        "title": "ParentPreferencesIn"
      },
...
      "PrestigeStability": {
        "type": "string",
        "enum": [
          "stability",
          "balanced",
          "prestige"
        ],
        "title": "PrestigeStability"
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "student",
          "parent",
          "educator",
          "admin"
        ],
        "title": "Role"
      },
...
    }
  }
}